JGNsc framework -- MCMC continuize function

JGNsc framework -- MCMC continuize function

JGNsc.zip(
  y,
  stepsize = 0.5,
  warm = 500,
  iter = 5000,
  min.cell = 1,
  a1 = 0.001,
  a2 = 0.001,
  b1 = 1e+08,
  b2 = 1e+08,
  a3 = 2.5,
  b3 = 1,
  c0 = 0
)

Arguments

y
matrix. dim: genes by samples

stepsize
the stepsize for the random proposal, default is 0.5

warm
the number of warm steps in the MCMC procedure

iter
the number of iterations in the MCMC procedure

min.cell
the min number of cells that should a gene should express

a1
the initial value of the parameter from the alpha parameter distribution: alpha ~ Gamma(a1,b1)

a2
the initial value of the parameter from the beta parameter distribution: beta ~ Gamma(a2,b2)

b1
the initial value of the parameter from the alpha parameter distribution: alpha ~ Gamma(a1,b1)

b2
the initial value of the parameter from the beta parameter distribution: beta ~ Gamma(a2,b2)

a3
the initial value of the parameter from the (non-dropout) p parameter distribution: p ~ Beta(a3,b3)

b3
the initial value of the parameter from the (non-dropout) p parameter distribution: p ~ Beta(a3,b3)

c0
a constant number close to zero, for the dropout case, theta value is imputed as c0 first.

Value

y.cont: imputed and continuized count matrix

keep.gene: genes kept