Matrix completion imputation method

the original algorithm was implemented in Matlab. Here, we rewrite it in R.

mcImpute.R(
  data,
  preprocess = T,
  eps = 1e-12,
  normfac = 1,
  insweep = 20,
  tol = 1e-04,
  decfac = 0.7,
  min_count = 2,
  min_cells = 3,
  verbose = F
)

Arguments

data
a matrix, samples by genes

preprocess
logical, whether to preprocess data: Removing BAD genes, median normalization and log-transformation.

eps
a small positive number that controls convergence, when calculating the Frobenius norm.

normfac
normalize factor, default is 1.

insweep
the max number of iterations when minimizing the rank of matrix given a tuning parameter(lambda) value

tol
a small positive number that controls convergence

decfac
decay factor, controls the rate when searching for the softthresholding parameter.

min_count
for filtering step, minimum number of counts that a gene expressed in a cell

min_cells
for filtering step, minimum number of cells that a gene expressed in.

verbose
logical, show message.

Source

https://github.com/aanchalMongia/McImpute_scRNAseq

Value

an imputed matrix

References

https://www.frontiersin.org/articles/10.3389/fgene.2019.00009/full