plot one network

plot one network

plot_onenet(
  parcorr,
  mode = "circle",
  gname = "",
  lsize = 2.5,
  nodecolor = "grey",
  family.vec = 0,
  circlenet = F
)

Arguments

parcorr
partial correlation matrix

mode
ggnet2 mode, could be "circle" or "spring",...

gname
character, graph name.

lsize
numeric, label size

nodecolor
vector/character, specify the color for each node

family.vec
vector, assign nodes to families. In other words, it could be the classification / covariates of the nodes.

circlenet
logical, whether plot the network in a circle shape.

Value

a network figure